Remove unnecessary (long) casts in MultiAnchorExecutor
Fix PMD `UnnecessaryCast` warnings at lines 174 and 176 by relying
on binary numeric promotion with `long` operands.

MARKDOWN=1
TAG=agy
CONV=3a3f1223-38da-46e5-a43f-07446bfafc13

+++ safere/src/main/java/org/safere/MultiAnchorExecutor.java
@@ -171,9 +171,8 @@ final class MultiAnchorExecutor {
             upstreamMatched = false;
             break;
           }
-          long maxAnchorBytes = (long) upstreamAnchor.maxLength() * 4L;
-          int firstOverlappingAnchorStart =
-              (int) Math.max(0L, (long) earliestGapStart - maxAnchorBytes);
+          long maxAnchorBytes = upstreamAnchor.maxLength() * 4L;
+          int firstOverlappingAnchorStart = (int) Math.max(0L, earliestGapStart - maxAnchorBytes);
           searchLowerBound = Math.max(searchLowerBound, firstOverlappingAnchorStart);
 
           if (searchUpperBound < searchLowerBound) {
